feat(arch): Make rayColor iterative not recursive
* Convert rayColor() to be iterative rather than recursive to facilitate
  easier parallelisation and be compatible with gpu hardware
  acceleration.
* Transition deprecated result_of<T> usage to use invoke_result<T>
* Change the method of displaying elapsed time in the Timer class to use
  fmtlib instead of native streams. This is because there is currently
  partial support for std::chrono operator<< - some versions of gcc
  require the operator defined in the date library because it's not
  present in libstdc++. Using fmtlib/chrono resolves these issues to
  make Timer compatible with more compilers.
* Add threadsafe compatibility to the randomNumber<T>() function.
* Remove parameter name for unised rInputRay parameter in
  Lambertian::scatter to silence compiler warnings.
* Cast RGB values to double when assigning to a Color to avoid
  narrowing compiler warnings in getRGB() function calls.

Closes #8.

+++ src/color.cpp
@@ -1,7 +1,7 @@
 #include "color.h"
+#include <cstdint>
 
 auto writeColor(Color rPixelColor) -> std::string {
-
     auto vRed = static_cast<uint8_t>(255.999 * rPixelColor.x());
     auto vGreen = static_cast<uint8_t>(255.999 * rPixelColor.y());
     auto vBlue = static_cast<uint8_t>(255.999 * rPixelColor.z());
@@ -25,7 +25,9 @@ auto getRGB(Color rPixelColor) -> Color {
     auto vRed = static_cast<uint8_t>(255.999 * rPixelColor.x());
     auto vGreen = static_cast<uint8_t>(255.999 * rPixelColor.y());
     auto vBlue = static_cast<uint8_t>(255.999 * rPixelColor.z());
-    return {vRed, vGreen, vBlue};
+    return {static_cast<double>(vRed),
+            static_cast<double>(vGreen),
+            static_cast<double>(vBlue)};
 };
 
 auto getRGB(Color rPixelColor, size_t rSamplesPerPixel) -> Color {
@@ -35,8 +37,10 @@ auto getRGB(Color rPixelColor, size_t rSamplesPerPixel) -> Color {
     auto vGreen = std::sqrt(vScale * rPixelColor.y());
     auto vBlue = std::sqrt(vScale * rPixelColor.z());
 
-    auto vScaledRed = static_cast<int>(256 * clamp(vRed, 0.0, 0.999));
-    auto vScaledGreen = static_cast<int>(256 * clamp(vGreen, 0.0, 0.999));
-    auto vScaledBlue = static_cast<int>(256 * clamp(vBlue, 0.0, 0.999));
-    return {vScaledRed, vScaledGreen, vScaledBlue};
+    auto vScaledRed = static_cast<uint8_t>(256 * clamp(vRed, 0.0, 0.999));
+    auto vScaledGreen = static_cast<uint8_t>(256 * clamp(vGreen, 0.0, 0.999));
+    auto vScaledBlue = static_cast<uint8_t>(256 * clamp(vBlue, 0.0, 0.999));
+    return {static_cast<double>(vScaledRed),
+            static_cast<double>(vScaledGreen),
+            static_cast<double>(vScaledBlue)};
 };
